Add file for utility functions.
